Оплата кредита

+++ create_table.sql
@@ -19,7 +19,7 @@ create table dbo.payment_list (
     _id                               integer identity(1,1) not null,
     _login                            integer not null,
     _date                             date not null,
-    _summ                             integer not null,
+    _summ                             float not null,
     constraint PK_users_list_table    primary key clustered(_id asc)
 );
